requirements.txt
click==8.1.3
cuid==0.4
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.7
sqlalchemy-views==0.3.2
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
